Feb 10, 2026, 6:58 PM

+++ main.cpp
@@ -55,7 +55,7 @@ int main()
 
   double shippingFee = getShippingCost(method, cost);
 
-  if (shippingFee == -1){
+  if (static_cast<int>(shippingFee) == -1){
     if (method == 1 && cost < 100){
       cout << "Not available";
     }
